Updated AIO/VFD branch from trunk and dealt with bugs and incompatibilities 
as necessary.  Amongst other things, this required deleting a number of files
unrelated to the AIO/VFD work as detailed below.

Modified test/vfd to be express test aware, and to issue notices when skipping
tests.

Added support for the H5_HAVE_AIO #define.  When H5_HAVE_AIO is not defined,
all but the top level AIO code is Eifdef'ed out.

With the help of Mike M, updated the configure code to define H5_HAVE_AIO 
for Linux and freebsd.  H5_HAVE_AIO is undefined in all other cases.

Tested serial and parallel on Phoenix.  Tested serial on jam, heiwa, linew, and
freedom.

+++ test/enum.c
@@ -71,7 +71,7 @@ test_named(hid_t file)
     if(H5Tclose(type) < 0) FAIL_STACK_ERROR
 
     /* A smaller type */
-    if((type = H5Tcreate(H5T_ENUM, 1)) < 0) FAIL_STACK_ERROR
+    if((type = H5Tcreate(H5T_ENUM, (size_t)1)) < 0) FAIL_STACK_ERROR
     if(H5Tenum_insert(type, "RED",   CPTR(val8, E1_RED  )) < 0) FAIL_STACK_ERROR
     if(H5Tenum_insert(type, "GREEN", CPTR(val8, E1_GREEN)) < 0) FAIL_STACK_ERROR
     if(H5Tenum_insert(type, "BLUE",  CPTR(val8, E1_BLUE )) < 0) FAIL_STACK_ERROR
